chore: split discovery files to packages
